// frontend/src/api/generated/model/gaiaMatchSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * astroimage
 * OpenAPI spec version: 0.1.0
 */
import type { GaiaMatchSchemaObjectType } from './gaiaMatchSchemaObjectType';

export interface GaiaMatchSchema {
  source_id: number;
  object_type: GaiaMatchSchemaObjectType;
  rank: number;
  ra_deg?: number | null;
  dec_deg?: number | null;
  gaia_match?: boolean;
  gaia_separation_arcsec?: number | null;
  gaia_probability?: number;
  gaia_source_id?: string | null;
  gaia_ra_deg?: number | null;
  gaia_dec_deg?: number | null;
  gaia_gmag?: number | null;
}